Add support for specifying the path inline including mode sigils
I ended up doing this on easy mode for now which results in the sigils prefixing the path and needed to be doubled up.

It will work and I'll be able to improve the parser down the line to make it so that the sigils do not have to be doubled-up.